Centralize random number generation and distance calculation logic
Consolidate the `mulberry32` random number generator into a single utility file and update four modules to import it. Also, centralize the `manhattan` distance calculation into a utility file, updating three modules to import it.
