+++ game/core/gm/runtime/threads/survey_cache.js
@@ -8,6 +8,7 @@
  */
 
 import { hash32 } from "../rng.js";
+import { mulberry32 } from "../../../../utils/rng.js";
 
 // ------------------------
 // Constants (mirrored from previous GMBridge implementation)
@@ -29,16 +30,7 @@ export const SURVEY_SALT_COOLDOWN = 0x534306;
 // Small deterministic helpers (copied semantics from gm_bridge.js)
 // ------------------------
 
-function mulberry32(seed) {
-  let a = (seed >>> 0);
-  return function () {
-    a = (a + 0x6d2b79f5) >>> 0;
-    let t = a;
-    t = Math.imul(t ^ (t >>> 15), t | 1);
-    t ^= t + Math.imul(t ^ (t >>> 7), t | 61);
-    return ((t ^ (t >>> 14)) >>> 0) / 4294967296;
-  };
-}
+// mulberry32 is imported from utils/rng.js
 
 function hash2Float(seed, x, y) {
   // Deterministic 2D hash -> [0,1)
