+++ game/world/infinite_gen.js
@@ -11,6 +11,7 @@
  * - Phase 1: sinusoidal climate temperature field and ridge-shaped mountains for the overworld.
  */
 import { attachGlobal } from "../utils/global.js";
+import { mulberry32 } from "../utils/rng.js";
 
 // Keep IDs aligned with World.TILES so renderers and logic work as-is
 const TILES = {
@@ -32,14 +33,7 @@ const TILES = {
   TOWER: 17,       // overworld tower POI; kept aligned with World.TILES
 };
 
-function mulberry32(a) {
-  return function () {
-    let t = (a += 0x6D2B79F5);
-    t = Math.imul(t ^ (t >>> 15), t | 1);
-    t ^= t + Math.imul(t ^ (t >>> 7), t | 61);
-    return ((t ^ (t >>> 14)) >>> 0) / 4294967296;
-  };
-}
+// mulberry32 is imported from utils/rng.js
 
 function hash2(seed, x, y) {
   // Deterministic 2D hash -> [0,1)
